legal-data-extractor/pyproject.toml
[project]
name = "legal-data-extractor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.85.0",
    "groq>=1.1.1",
    "kagglehub>=1.0.0",
    "pdfplumber>=0.11.9",
    "python-dotenv>=1.2.2",
    "tqdm>=4.67.3",
    "transformers>=5.3.0",
]
